Add CombatEngine and delegate combat logic
